Updates to the database schema and some more fields added

+++ 9_IMS_Final/Products.cs
@@ -4,6 +4,6 @@
     {
         public string Prod_ID { get; set; }
         public string Prod_Name { get; set; }
-        public float Re_Order { get; set; }
+        public int Re_Order { get; set; }
     }
 }
